backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
supabase==2.11.0
cryptography==44.0.0
apscheduler==3.10.4
openai==1.59.6
PyGithub==2.5.0
httpx==0.28.1
python-dotenv==1.0.1
email-validator==2.1.1
